move non-conflicting, non-application assets from konsole to kpart package
this includes:
- kpart service file (kservices5)
- kpart terminal emulator type (kservicetypes5)
- keyboard layouts & colors (konsole)
- notification profiles (knotifications5)
- ui definitions (kxmlgui5)

out of those only one of the 3 ui definitions isn't used by the kpart, so
this for the most part only moves all relevant bits into the kpart package.

generally speaking we should move forward packaging all semi-versioned bits
into the kpart package to make sure the kpart is actually functional.

NB: share/konsole/ is not conflicting because in kde4 we had
    share/kde4/konsole, this is however something to watch out for come
    qt6. we should push for proper compatibility versioning of there
    that folder then.